Create Finance bot
This code defines a Telegram bot that responds to the /start command with a welcome message and the /stock <symbol> command with the current stock price of the specified symbol. It fetches stock prices using the Alpha Vantage API and sends them back to the user. Make sure to replace "YOUR_TELEGRAM_BOT_TOKEN" and "YOUR_ALPHA_VANTAGE_API_KEY" with your actual tokens and credentials.